Extra read-only superclass for LongColumnSourceRedirectionIndex
